+++ CHANGELOG.md
@@ -1,4 +1,4 @@
 # Changelog
 
 ## V8.0
-- Initial Dependency Injection Engine.
+- Initial API Engine.